requirements.txt
streamlit
sentence-transformers
faiss-cpu
transformers
torch
numpy
langchain
langchain-text-splitters
openai